修改ds3231驱动

+++ component/edriver/DS3231.cpp
@@ -42,12 +42,12 @@ uint8_t DS3231::dec_to_bcd(uint8_t dec)
 
     return bcd;
 }
-
+//uint8_t slaveAddr, uint8_t regAddr, uint8_t *data, uint16_t nRead
 void	DS3231::get_date_time(DateTime_t *t)
 {
     uint8_t buf[8];
     i2c->take(speed);
-    i2c->read_buf(DS3231_ADDRESS, DS3231_SECOND, buf, 7);
+    i2c->read_buf(DS3231_ADDRESS, (uint8_t)DS3231_SECOND, buf, 7);
     i2c->release();
 
     //	timer.w_year,timer.w_month,timer.w_date,timer.hour,timer.min,timer.sec
@@ -65,7 +65,7 @@ void DS3231::get_date(char *buf)
     uint8_t tmpbuf[3];
     i2c->take(speed);
 
-    i2c->read_buf(DS3231_ADDRESS, DS3231_DAY, tmpbuf, 3);				//日期
+    i2c->read_buf(DS3231_ADDRESS, (uint8_t)DS3231_DAY, tmpbuf, 3);				//日期
     i2c->release();
 
     buf[0] = uint8_t( (tmpbuf[0] >> 4) + 0x30);
@@ -81,7 +81,7 @@ void DS3231::get_time(char *buf)
     uint8_t tmpbuf[3];
 
     i2c->take(speed);
-    i2c->read_buf(DS3231_ADDRESS, DS3231_SECOND, tmpbuf, 3);
+    i2c->read_buf(DS3231_ADDRESS, (uint8_t)DS3231_SECOND, tmpbuf, 3);
     i2c->release();
 
     buf[0] = uint8_t( (tmpbuf[2] >> 4) + 0x30);
@@ -109,13 +109,13 @@ void	DS3231::set_time(void *dt)
 
 
     i2c->take(speed);
-    i2c->write(DS3231_ADDRESS, DS3231_WEEK, tBCD.week); //修改周
-    i2c->write(DS3231_ADDRESS, DS3231_YEAR, tBCD.year); //修改年
-    i2c->write(DS3231_ADDRESS, DS3231_MONTH, tBCD.month); //修改月
-    i2c->write(DS3231_ADDRESS, DS3231_DAY, tBCD.date);  //修改日
-    i2c->write(DS3231_ADDRESS, DS3231_HOUR, tBCD.hour); //修改时
-    i2c->write(DS3231_ADDRESS, DS3231_MINUTE, tBCD.min); //修改分
-    i2c->write(DS3231_ADDRESS, DS3231_SECOND, tBCD.sec ); //修改秒
+    i2c->write(DS3231_ADDRESS, (uint8_t)DS3231_WEEK, tBCD.week); //修改周
+    i2c->write(DS3231_ADDRESS, (uint8_t)DS3231_YEAR, tBCD.year); //修改年
+    i2c->write(DS3231_ADDRESS, (uint8_t)DS3231_MONTH, tBCD.month); //修改月
+    i2c->write(DS3231_ADDRESS, (uint8_t)DS3231_DAY, tBCD.date);  //修改日
+    i2c->write(DS3231_ADDRESS, (uint8_t)DS3231_HOUR, tBCD.hour); //修改时
+    i2c->write(DS3231_ADDRESS, (uint8_t)DS3231_MINUTE, tBCD.min); //修改分
+    i2c->write(DS3231_ADDRESS, (uint8_t)DS3231_SECOND, tBCD.sec ); //修改秒
     i2c->release();
 }
